+++ server/database/users.py
@@ -2,7 +2,7 @@ from typing import Optional
 
 import mariadb
 
-from models import User
+from .models import (User)
 from .utils import generate_uid